Update to G_H function
This is a more stable fit of Gamma(z) in log than in linear space, if you remember the  Jacob Seiler's emails (will reply to it now).
Problem was the fit in linear goes to the negative values of Gamma, which in turn gives -NAN when applying XHI(ratio) function due to having negative in the square root. I made a modification to the parameters when the fit is in log space. This should work with no problems. You may direct Jacob to use this updated function, I am very sorry for this, I should have noticed this, but this bug only was affecting some redshifts between z=11 to z=10. Sultan